repo: make `clippy --workspace -D warnings` and `biome check` green
The two lint gates have never passed on this repo, so nothing new could
be held to them either — `cargo clippy -- -D warnings` applies to every
workspace path dependency, and `pnpm lint` fails the whole tree. Both
are clean now, with no behaviour change anywhere.

clippy (2058 -> 0):
  * 2054 of them were `gl_bindings_generated.rs`, which build.rs
    regenerates from gl_generator on every clean build — there is no
    source file to fix, so the three lints its output shape trips are
    allowed at the `include!` site alongside the rustc lints already
    there.
  * the remaining 16 are real and fixed as findings, not silenced:
    `attach_window_node`'s five positional f64s become a `NodeGeometry`
    struct (transposing width/height was previously invisible),
    `AppState`'s three nested generics become `WindowKey` /
    `FramesByClient` aliases, `XkbState`'s hand-written `Default` is
    derived, an `x = Default::default(); x.f = ...` becomes struct
    update syntax, plus a needless borrow, an identity cast and a
    doc-comment line that started with `+` and so parsed as a list.
  * also clears the two long-standing rustc warnings: a genuinely dead
    `let prev`, and a `ReplyBuf` import now cfg'd to the linux-only
    CreateSegment path that uses it.

biome (42 errors -> 0), all pre-existing frontend findings:
  * a11y is fixed, not suppressed. Polaroid becomes a real `<button>`
    (it was a div with onClick, so unreachable by keyboard); toolbar
    glyphs get `aria-hidden`; the canvas viewport, window frame and
    markdown card get `role="application"`, which is what they are;
    MarkdownArea's `role="textbox"` gets the accessible name axe asks
    for, via a new optional `label` prop.
  * the five `useExhaustiveDependencies` hits in canvasInputHooks were
    carrying `eslint-disable` comments that have been inert since the
    biome migration; they are now real biome suppressions naming the
    reason (RefObject *parameters* — `.current` is not reactive).
  * `corner-shape: squircle` is valid CSS that biome's property table
    predates; noArrayIndexKey in MarkdownArea is load-bearing (the
    caret's span must keep its identity). Both suppressed with why.
  * xts.spec's `while ((m = re.exec(s)))` becomes `matchAll`.

`.claude` is excluded from biome: it is Claude Code's local state, not
project source, and its presence made the pre-commit hook fail.

Verified: cargo check/test --workspace (host), wayland-build.sh
test/clippy --workspace (linux), both sidecar images, turbo typecheck,
biome check, wayland-smoke.sh, and every JS unit suite (components
68/68, frontend 116/116, canvas-core 20/20, canvas-r3f 6/6).

+++ crates/x11-server/src/xserver/handlers/window/attributes.rs
@@ -298,7 +298,7 @@ pub(crate) fn handle_get_window_attributes(
             colormap: cmap,
             all_event_masks: EventMask::from(win.event_mask | remote_masks),
             your_event_mask: EventMask::from(win.event_mask),
-            do_not_propagate_mask: EventMask::from(win.do_not_propagate_mask as u32),
+            do_not_propagate_mask: EventMask::from(win.do_not_propagate_mask),
         },
         state.byte_order(),
     )
